Resync from spuce with fix for Sinc filter, use cpp files for design_, add new classes/files, update library name to libspuce, update FIRDesigner for Sinc case, change beta to alpha for raised cosine filters, dont allow Maxflat to be used for bandstop filters

+++ comms/spuce/filters/transform_fir.cpp
@@ -29,7 +29,7 @@ std::vector<float_type> transform_fir(const std::string& band_type, const std::v
   } else if (band_type == "BAND_STOP") {
 		// First transform to band_pass, then subtract 1 from center tap
 		out = transform_fir("BAND_PASS", taps, f);
-		int mid_tap = (taps.size()-1)/2;
+		size_t mid_tap = (taps.size()-1)/2;
     out[mid_tap] -= 1;
   } else {
     std::cout << "Unsupported band type :" << band_type << "\n";
@@ -39,14 +39,14 @@ std::vector<float_type> transform_fir(const std::string& band_type, const std::v
 std::vector<std::complex<float_type>> transform_complex_fir(const std::string& band_type, const std::vector<float_type>& taps, float_type f) {
   std::vector<std::complex<float_type>> out(taps.size());
   if (band_type == "COMPLEX_BAND_PASS") {
-		// Modulate taps up to band center frequency
-		for (size_t i=0;i<taps.size();i++) {
-			double phase = 2.0*M_PI*f*i;
-			out[i] = taps[i]*std::polar(1.0, phase);
-		}
+    // Modulate taps up to band center frequency
+    for (size_t i=0;i<taps.size();i++) {
+      double phase = 2.0*M_PI*f*i;
+      out[i] = taps[i]*std::polar(1.0, phase);
+    }
   } else if (band_type == "COMPLEX_BAND_STOP") {
 		out = transform_complex_fir("COMPLEX_BAND_PASS", taps, f);
-		int mid_tap = (taps.size()-1)/2;
+		size_t mid_tap = (taps.size()-1)/2;
 		out[mid_tap] -= 1;
   } else {
     std::cout << "Unsupported band type :" << band_type << "\n";
